moved apiCalls to utilities
